refactor: 💡 重構 modal 來重用
